# Remove the 'close' and 'opened' fields.